+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Latest
-
+- Add `benchmark-pipeline` command to run benchmarks.
 - The `read_model` stage can accept a FITS spectral cube as a model image.
 - The `read_model` and `imaging` stage parameters have changed. Please refer the "Stage Configs" section of the documentation.
 - The `cont_sub` stage reports the peak channel, and is controlled via the configuration parameter `report_peak_channel` of type boolean.